% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.gillnet}
\alias{AIC.gillnet}
\title{AIC of gillnet object}
\usage{
\method{AIC}{gillnet}(object, ...)
}
\arguments{
\item{object}{gillnet fitted object (result from gillnetfitTMB)}

\item{...}{extra arguments}
}
\description{
AIC of gillnet object
}
\details{
...
}
