% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.gillnetset}
\alias{AIC.gillnetset}
\title{AIC of gillnetset object}
\usage{
\method{AIC}{gillnetset}(object, ...)
}
\arguments{
\item{object}{gillnet fitted object (result from gillnetfitTMB)}

\item{...}{extra arguments}
}
\description{
AIC of gillnetset object
}
\details{
...
}
