% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.gillnet}
\alias{c.gillnet}
\title{Collect gillnet objects}
\usage{
\method{c}{gillnet}(...)
}
\arguments{
\item{...}{gillnet fits to be combined}
}
\description{
Collect gillnet objects
}
\details{
...
}
