% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.gillnet}
\alias{coef.gillnet}
\title{Extract fixed selectivity coefficients of gillnet object}
\usage{
\method{coef}{gillnet}(object, ...)
}
\arguments{
\item{object}{gillnet fitted object (result from illnetfitTMB)}

\item{...}{extra arguments}
}
\description{
Extract fixed selectivity coefficients of gillnet object
}
\details{
...
}
