% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{defdat}
\alias{defdat}
\title{Data validation}
\usage{
defdat(x, warning = FALSE)
}
\arguments{
\item{x}{input data (list)}

\item{warning}{logical}
}
\description{
Data validation
}
\details{
Transforms data input, for easy integration in TMB. Lengths are removed for which only zeros were observed across all mesh sizes, years, regions and periods (see Surette).
}
\examples{
x <- defdat(data)
}
