% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fittable}
\alias{fittable}
\alias{fittable.gillnet}
\alias{fittable.gillnetset}
\title{fittable}
\usage{
fittable(x, ...)

\method{fittable}{gillnet}(x)

\method{fittable}{gillnetset}(x)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}
}
\description{
fittable
}
\details{
...
}
