% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{gillnetR}
\alias{gillnetR}
\title{Fit gillnet selectivity model}
\usage{
gillnetR(par, x)
}
\arguments{
\item{par}{parms}

\item{x}{data}
}
\description{
Fit gillnet selectivity model
}
\details{
Gillnet model code (R version). Used to validate initial parameter estimation (defpar). Note: does not compile with RTMB (complex number conversion).
}
