% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.gillnet}
\alias{logLik.gillnet}
\title{Log likelihood of gillnet object}
\usage{
\method{logLik}{gillnet}(object, ...)
}
\arguments{
\item{object}{gillnet fitted object (result from gillnetfitTMB)}

\item{...}{extra arguments}
}
\description{
Log likelihood of gillnet object
}
\details{
...
}
