% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{Ntable}
\alias{Ntable}
\alias{Ntable.gillnet}
\alias{Ntable.gillnetset}
\title{Ntable}
\usage{
Ntable(x, ...)

\method{Ntable}{gillnet}(x)

\method{Ntable}{gillnetset}(x)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}
}
\description{
Ntable
}
\details{
...
}
