% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{partable}
\alias{partable}
\alias{partable.gillnet}
\alias{partable.gillnetset}
\title{partable}
\usage{
partable(x, ...)

\method{partable}{gillnet}(x, ...)

\method{partable}{gillnetset}(x, ...)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}
}
\description{
partable
}
\details{
...
}
