% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotN}
\alias{plotN}
\alias{plotN.gillnet}
\alias{plotN.gillnetset}
\title{plotN}
\usage{
plotN(x, ...)

\method{plotN}{gillnet}(x, ylab = "Relative abundance index", xlab = "Length")

\method{plotN}{gillnet}(
  x,
  ylab = "Relative abundance index",
  xlab = "Length",
  filllab = "Fit",
  dodge = 1.8
)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{ylab}{ylab}

\item{xlab}{xlab}

\item{x...}{}
}
\description{
plotN
}
\details{
...
}
