% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOP}
\alias{plotOP}
\alias{plotOP.gillnet}
\alias{plotOP.gillnetset}
\title{plot observed vs predicted}
\usage{
plotOP(x, ...)

\method{plotOP}{gillnet}(x, log = FALSE, ...)

\method{plotOP}{gillnet}(x, log = FALSE, ...)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}

\item{ylab}{ylab}

\item{xlab}{xlab}
}
\description{
plot observed vs predicted
}
\details{
...
}
