% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRes}
\alias{plotRes}
\alias{plotRes.gillnet}
\alias{plotRes.gillnetset}
\title{plotRes}
\usage{
plotRes(x, ...)

\method{plotRes}{gillnet}(x, xlab = "Length", ylab = "Mesh size")

\method{plotRes}{gillnet}(x, ylab = "Selectivity", xlab = "Length", collab = "Mesh")
}
\arguments{
\item{...}{extra arguments not currently used}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{x...}{}
}
\description{
plotRes
}
\details{
...
}
