% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSel}
\alias{plotSel}
\alias{plotSel.gillnet}
\alias{plotSel.gillnetset}
\title{plotSel}
\usage{
plotSel(x, ...)

\method{plotSel}{gillnet}(
  x,
  ylab = "Selectivity",
  xlab = "Length",
  collab = "Mesh",
  meshlabs = NULL
)

\method{plotSel}{gillnet}(
  x,
  ylab = "Selectivity",
  xlab = "Length",
  collab = "Mesh",
  meshlabs = NULL
)
}
\arguments{
\item{x}{gillnetset object}

\item{...}{extra arguments not currently used}

\item{ylab}{ylab}

\item{xlab}{xlab}

\item{collab}{color lab}

\item{meshlabs}{color labels}

\item{x...}{}
}
\description{
plotSel
}
\details{
...
}
