% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.gillnet}
\alias{predict.gillnet}
\title{Predict selectivity}
\usage{
\method{predict}{gillnet}(
  object,
  length = seq(min(object$data$length), max(object$data$length), by = 0.01),
  ...
)
}
\arguments{
\item{object}{gillnet fit object}

\item{...}{}
}
\description{
Predict selectivity
}
\details{
...
}
