% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.gillnetset}
\alias{predict.gillnetset}
\title{Predict selectivity}
\usage{
\method{predict}{gillnetset}(object, ...)
}
\arguments{
\item{object}{gillnetset object}

\item{...}{}
}
\description{
Predict selectivity
}
\details{
...
}
