% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predSel}
\alias{predSel}
\title{Predict selectivity}
\usage{
predSel(length, mesh, k1, k2, rtype = "norm.loc")
}
\arguments{
\item{length}{vector}

\item{mesh}{vector}

\item{k1}{param}

\item{k2}{param}

\item{rtype}{type of selectivity curve}
}
\description{
Predict selectivity
}
\details{
Predicts selectivity based on knowm parameter values
}
