% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{predtable}
\alias{predtable}
\alias{predtable.gillnet}
\alias{predtable.gillnetset}
\title{predtable}
\usage{
predtable(x, ...)

\method{predtable}{gillnet}(x)

\method{predtable}{gillnetset}(x)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}
}
\description{
predtable
}
\details{
...
}
