% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gillnet}
\alias{print.gillnet}
\title{Print gillnet object}
\usage{
\method{print}{gillnet}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print gillnet object
}
\details{
...
}
