% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gillnetcoef}
\alias{print.gillnetcoef}
\title{Print gillnetcoef object}
\usage{
\method{print}{gillnetcoef}(x, ...)
}
\arguments{
\item{x}{...}

\item{...}{extra arguments}
}
\description{
Print gillnetcoef object
}
\details{
...
}
