% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.gillnet}
\alias{residuals.gillnet}
\title{Extract residuals from gillnet object}
\usage{
\method{residuals}{gillnet}(object)
}
\arguments{
\item{object}{gillnet fitted object (result from gillnet.fit)}
}
\description{
Extract residuals from gillnet object
}
\details{
...
}
