% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.gillnetset}
\alias{residuals.gillnetset}
\title{Extract residuals from gillnetset object}
\usage{
\method{residuals}{gillnetset}(object)
}
\arguments{
\item{object}{gillnetset fitted objects (results from gillnet.fit)}
}
\description{
Extract residuals from gillnetset object
}
\details{
...
}
