% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{selmaxtable}
\alias{selmaxtable}
\alias{selmaxtable.gillnet}
\alias{selmaxtable.gillnetset}
\title{selmaxtable}
\usage{
selmaxtable(x, ...)

\method{selmaxtable}{gillnet}(x)

\method{selmaxtable}{gillnetset}(x)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{gillnet or gillnetset object}
}
\description{
selmaxtable
}
\details{
...
}
