% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{seltable}
\alias{seltable}
\alias{seltable.gillnet}
\alias{seltable.gillnetset}
\title{seltable}
\usage{
seltable(x, ...)

\method{seltable}{gillnet}(x)

\method{seltable}{gillnetset}(x)
}
\arguments{
\item{...}{extra arguments not currently used}

\item{x...}{}
}
\description{
seltable
}
\details{
...
}
