% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{defpar}
\alias{defpar}
\title{Default parameters}
\usage{
defpar(x, plot = FALSE)
}
\arguments{
\item{x}{input data (list)}

\item{plot}{plot selectivity curves at inital parameter estimates.}
}
\description{
Default parameters
}
\details{
Defaut parameters for gillnet selectivity function. Based on code Surette 2016.
mu initiation: In logsel = -(length-k1 x mesh)^2/(2 x k2^2), at peak selectivity (logSel=0), the equation can be solved so k1 = length/mesh.
}
\examples{
x <- defpar(data)
}
