% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotOP}
\alias{plotOP}
\alias{plotOP.gillnet}
\alias{plotOP.gillnetset}
\title{plot observed vs predicted}
\usage{
plotOP(x, ...)

\method{plotOP}{gillnet}(
  x,
  log = FALSE,
  xlab = "Observed",
  ylab = "Predicted",
  collab = "Mesh",
  facet = FALSE,
  meshlabs = NULL,
  ...
)

\method{plotOP}{gillnetset}(
  x,
  log = FALSE,
  xlab = "Observed",
  ylab = "Predicted",
  collab = "Mesh",
  ...
)
}
\arguments{
\item{x}{gillnetset object}

\item{...}{extra arguments not currently used}

\item{log}{logical}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{facet}{plot mesh as facet}

\item{meshlabs}{labels for mesh facets (if facet=TRUE)}

\item{x...}{}

\item{collabl}{color lab}
}
\description{
plot observed vs predicted
}
\details{
...
}
